+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This CHANGELOG is automatically compiled.
 
 #### ValidationKit
 
+- added Validate::uuid() ([c3f965f](https://github.com/stuartherbert/stusdevkit/commit/c3f965f)) — Stuart Herbert <stuart@stuartherbert.com>
 - added Validate::when() ([a4e5e02](https://github.com/stuartherbert/stusdevkit/commit/a4e5e02)) — Stuart Herbert <stuart@stuartherbert.com>
 - added missing JSON schema logic/sub-schema support ([5950061](https://github.com/stuartherbert/stusdevkit/commit/5950061)) — Stuart Herbert <stuart@stuartherbert.com>
 - add Validate::nullish() ([f3e7be5](https://github.com/stuartherbert/stusdevkit/commit/f3e7be5)) — Stuart Herbert <stuart@stuartherbert.com>
@@ -45,7 +46,7 @@ This CHANGELOG is automatically compiled.
 
 #### ValidationKit
 
-- make it easier for third-party validators to create validation issues ([b4c206b](https://github.com/stuartherbert/stusdevkit/commit/b4c206b)) — Stuart Herbert <stuart@stuartherbert.com>
+- make it easier for third-party validators to create validation issues ([5a7b06d](https://github.com/stuartherbert/stusdevkit/commit/5a7b06d)) — Stuart Herbert <stuart@stuartherbert.com>
 - Validate::discriminatedAnyof() performance improvement ([204b3af](https://github.com/stuartherbert/stusdevkit/commit/204b3af)) — Stuart Herbert <stuart@stuartherbert.com>
 
 #### AssertionsKit